+++ plugin/src/main/groovy/org/openbakery/signing/ProvisioningProfileReader.groovy
@@ -167,7 +167,11 @@ class ProvisioningProfileReader {
 		setBundleIndentiferToEntitlementsForValue(entitlementFile, bundleIdentifier, "application-identifier")
 		setBundleIndentiferToEntitlementsForValue(entitlementFile, bundleIdentifier, "com.apple.application-identifier")
 		setBundleIndentiferToEntitlementsForValue(entitlementFile, bundleIdentifier, "com.apple.developer.ubiquity-kvstore-identifier")
-		setBundleIndentiferToEntitlementsForValue(entitlementFile, bundleIdentifier, "keychain-access-groups")
+		//setBundleIndentiferToEntitlementsForValue(entitlementFile, bundleIdentifier, "keychain-access-groups")
+
+		// the keychain-access-groups were removed from the entitlements on the xcode export step
+		// therefor I also remove this here. Maybe this is wrong, but I hope the future will clarify this.
+		plistHelper.deleteValueFromPlist(entitlementFile, "keychain-access-groups")